+++ dbconsole_update_worker.py
@@ -432,6 +432,7 @@ class UpdateWorker:
             "LOCAL_MYSQL_PROFILE_NAME",
             "LOCAL_MYSQL_ADMIN_USER",
             "LOCAL_MYSQL_ADMIN_PASSWORD",
+            "LOCAL_MYSQL_ROOT_PASSWORD",
             "EMBEDDED_MYSQL_SHELL_DIR",
             "EMBEDDED_MYSQL_SERVER_DIR",
             "MYSQL_SERVER_VERSION",